wallet/requirements.txt
amqp==5.3.1
asgiref==3.11.1
async-timeout==5.0.1
billiard==4.2.4
celery==5.6.2
celery-types==0.24.0
certifi==2026.1.4
charset-normalizer==2.1.1
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==3.2
djangorestframework==3.14.0
exceptiongroup==1.3.1
Flask==2.1.3
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.6.2
MarkupSafe==3.0.3
packaging==26.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
pytz==2025.2
redis==7.1.1
requests==2.28.1
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==1.26.20
vine==5.1.0
wcwidth==0.6.0
Werkzeug==2.0.3
